Moved Editor into separate component
